requirements.txt
openai
pixellab
sounddevice
numpy
scipy
Pillow
python-dotenv
